feat: Add yoga class management
This feature introduces the ability to create, view, edit, and delete individual yoga classes associated with a yoga course.

Key changes:
- **Class Management:**
  - Added CreateClassActivity and YogaClassListActivity to manage yoga classes.
  - Implemented functionality to create, edit, delete, and update the status of classes.
- **Database & Models:**
  - Created a new YogaClass entity, DAO, and repository.
  - Updated the AppDatabase to include the new YogaClass table and incremented the database version.
  - Changed the id type from int to long in YogaCourse and related files for consistency.
- **UI & Navigation:**
  - Updated MainActivity to navigate to the class list when a course is clicked.
  - Added new layouts and drawables to support the class management feature.
- **Dependencies:**
  - Added lifecycle-viewmodel and lifecycle-livedata to support the new ViewModels.

+++ admin-app/app/src/main/java/com/example/yogaAdmin/activities/CreateCourseActivity.java
@@ -41,7 +41,7 @@ public class CreateCourseActivity extends AppCompatActivity {
     // Error icons for required fields
     private ImageView ivErrorDayOfWeek, ivErrorTime, ivErrorCapacity, ivErrorDuration, ivErrorPrice, ivErrorClassType;
 
-    private int courseId = -1;
+    private long courseId = -1;
     private long existingCourseCreatedDate = 0;
 
     @Override
@@ -357,3 +357,4 @@ public class CreateCourseActivity extends AppCompatActivity {
         });
     }
 }
+
